+++ project_1/include/NetworkSimulator.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "server.hpp"
-#include "event.hpp"
-#include "packet.hpp"
+#include "ServerSender.hpp"
+#include "Event.hpp"
+#include "PacketHeader.hpp"
 
 #include <vector>
 #include <cstdint>
@@ -16,22 +16,15 @@ public:
 
     void StartSimulation(std::uint32_t simulation_time_sec);
 
-    // New ID = last given ID + 1
-    template<typename T>
-    std::uint64_t GenerateNewID(const std::vector<T>& data) {
-        std::uint64_t new_id = 1;
-        if (!data.empty()) {
-            new_id = data.back().GetID() + 1;
-        }
-        return new_id;
-    }
-
     friend std::ostream& operator<<(std::ostream& out, const NetworkSimulator& simulator);
 private:
     const std::uint64_t bandwidth_bytes = 6'250'000'000;
     const std::uint64_t data_transmission_frequency_bytes = 6'250'000'000;
     const std::uint64_t data_transmission_frequency_packets = 6'250'000;
     const std::uint32_t packet_size_bytes = 1024;
+
+    // New ID = last given ID + 1
+    std::uint64_t last_given_server_id = 0;
 
     // Contains all sending servers 
     std::vector<ServerSender> servers{};
