+++ project_1/include/ServerSender.hpp
@@ -1,25 +1,12 @@
 #pragma once
 
-#include "packet.hpp"
-
-#include <cstdint>
-#include <vector>
-#include <queue>
-
-// Type for server receiver, contains only ID
-struct ServerBase {
-    ServerBase(std::uint32_t id);
-    ServerBase() = default;
-    virtual ~ServerBase() = default;
-
-    std::uint32_t GetID() const;
-protected:
-    std::uint64_t id;
-};
+#include "ServerBase.hpp"
+#include "ACK.hpp"
+#include "PacketHeader.hpp"
 
 // Type for sending server, contains ID and distance to receiver
 struct ServerSender : ServerBase {
-    ServerSender(std::uint32_t id, std::uint32_t distance_us, std::priority_queue<PacketHeader>& queue);
+    ServerSender(std::uint64_t id, std::uint32_t distance_us);
     ServerSender() = default;
     virtual ~ServerSender() = default;
 
@@ -34,13 +21,12 @@ struct ServerSender : ServerBase {
     void AddACK(const ACK& ack);
     bool HandleACKs();
     void printACKs() const;
-    void SendPackets();
+    void SendPackets(std::priority_queue<PacketHeader>& packet_queue);
 private:
     const std::uint32_t max_transmission_rate_packets = 6'250'000;
     const std::uint32_t distance_us;
     std::uint32_t current_time_us = 0;
     std::uint32_t cwnd_size_in_packets = 1;
 
-    std::priority_queue<PacketHeader>& packet_queue;
     std::vector<ACK> ACKs{};
 };
