+++ project_1/src/main.cpp
@@ -1,4 +1,4 @@
-#include "simulator.hpp"
+#include "NetworkSimulator.hpp"
 
 #include <iostream>
